pyproject.toml
[project]
name = "langgraph-studio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub>=0.29.3",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.2.3",
    "streamlit>=1.43.2",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.1.77",
    "watchdog>=6.0.0",
]
